`migrate` command CLI finished & tested
added optional `custom_usage` to click class factory
added optional `quiet` to `set_log_level` func
